+++ packages/ctl/src/command_ops.ts
@@ -1,0 +1,4 @@
+export function print(...content: any) {
+	// eslint-disable-next-line no-console
+	console.log(...content);
+}
